+++ deepinv/loss/score.py
@@ -1,5 +1,4 @@
 import torch
-import torch.nn as nn
 import numpy as np
 from deepinv.loss.loss import Loss