+++ deepinv/utils/nn.py
@@ -63,14 +63,6 @@ class TensorList:
             )
         return self
 
-    def conj(self):
-        r"""
-        Returns the complex conjugate of the list of tensors.
-
-        If the tensors are real, it returns the same list.
-        """
-        return TensorList([xi.conj() for xi in self.x])
-
     def __add__(self, other):
         r"""
 
@@ -257,24 +249,6 @@ def get_freer_gpu(verbose=True):
         print("Couldn't find free GPU")
 
     return device
-
-
-def save_model(
-    epoch, model, optimizer, ckp_interval, epochs, loss, save_path, eval_psnr=None
-):
-    if (epoch > 0 and epoch % ckp_interval == 0) or epoch + 1 == epochs:
-        os.makedirs(save_path, exist_ok=True)
-
-        state = {
-            "epoch": epoch,
-            "state_dict": model.state_dict(),
-            "loss": loss,
-            "optimizer": optimizer.state_dict(),
-        }
-        if eval_psnr is not None:
-            state["eval_psnr"] = eval_psnr
-        torch.save(state, os.path.join(save_path, "ckp_{}.pth.tar".format(epoch)))
-    pass
 
 
 def load_checkpoint(model, path_checkpoint, device):